+++ bonfire/docs/changelog.md
@@ -7,6 +7,7 @@
 #### New Features:
 
 #### Closes Issues:
+* #1129 Module Builder: filenames and class names built with incorrect case for CI3.
 * #1128 Module Builder: Use of `strip_slashes()` without loading the string helper.
 * #1118 Settings error when password options are not selected in security tab.